Optimized Golem Creation/Worked on NavMesh of the Golems
The answer for the golem creation optimization was already implemented, I just didn't realize it lol. All I had to do was instead of checking for if the piece's scrap pile's name equaled that of the scrap object last saved in animate point, to instead set the piece's scrap pile to equal that of the scrap pile itself :D. More Optimization is to come, BUT I am going to start working on the nav meshes of the golem, which will also eventually translate to the scrap wolves.