It's all basically done
it turns out it wasn't the menu stuff, it was when the pile was being assigned at the on trigger enter, when it should have been set to on trigger stay in the GolemCreation thing. The only thing left is to create a reference to the UI selector in the gauntlet script for the gameover functions and the pause menu.